+++ bluetooth_low_energy_windows/lib/src/central_manager_impl.dart
@@ -368,6 +368,7 @@ final class CentralManagerImpl
     final state = stateArgs.toState();
     final eventArgs = PeripheralConnectionStateChangedEventArgs(
       peripheral,
+      0,
       state,
     );
     _connectionStateChangedController.add(eventArgs);